requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
chardet==4.0.0
click==8.1.7
dataclasses==0.6
deprecation==2.1.0
Flask==2.3.2
gotrue==0.2.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
postgrest-py==0.4.0
py==1.11.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
supabase-py==0.0.2
toml==0.10.2
urllib3==1.26.16
websockets==9.1
Werkzeug==2.3.7
